correct logic in brute force #best_pair
Just above, we say

> A natural choice might be to treat the profiles as points in an 3-dimensional space; the "similarity" could be -1 times the distance between the two points.

On that basis, the best-matched cats (the most similar cats, the ones nearest to each other in the three dimensional space) will have the highest (that is, least negative) similarity score. So we want to skip to the next cat pair with `next` iff we have a best_similarity and that best_similarity is better than (that is, *greater than*) the other similarity currently under consideration. Otherwise, the new similarity is better than (that is, greater than) the old best_similarity, so we want to proceed to the next line where we assign the new pair and similarity measure to the best_pair and best_similarity variables.